# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/bootstrap": {
      "get": {
        "operationId": "getOwnerBootstrap",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Load owner Mini App bootstrap data",
        "tags": [
          "bootstrap"
        ]
      }
    },
    "/v1/matches": {
      "get": {
        "operationId": "listOwnerMatches",
        "parameters": [
          {
            "name": "venueId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "List current information cards",
        "tags": [
          "matches"
        ]
      },
      "post": {
        "operationId": "createOwnerMatch",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Create and publish an information card",
        "tags": [
          "matches"
        ]
      }
    },
    "/v1/matches/{id}": {
      "get": {
        "operationId": "getOwnerMatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Load one information card",
        "tags": [
          "matches"
        ]
      },
      "patch": {
        "operationId": "updateOwnerMatch",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Update an information card",
        "tags": [
          "matches"
        ]
      },
      "delete": {
        "operationId": "deleteOwnerMatch",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true,
                    "matchId": "42"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Delete an information card",
        "tags": [
          "matches"
        ]
      }
    },
    "/v1/weather/current": {
      "post": {
        "operationId": "sendCurrentWeather",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherCurrentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Send current Minsk weather to Telegram",
        "tags": [
          "weather"
        ]
      }
    },
    "/v1/venues": {
      "get": {
        "operationId": "listOwnerVenues",
        "parameters": [
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "List venue catalog entries",
        "tags": [
          "venues"
        ]
      },
      "post": {
        "operationId": "createOwnerVenue",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Create a venue catalog entry",
        "tags": [
          "venues"
        ]
      }
    },
    "/v1/venues/{id}": {
      "patch": {
        "operationId": "updateOwnerVenue",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Update a venue catalog entry",
        "tags": [
          "venues"
        ]
      }
    },
    "/v1/venues/{id}/archive": {
      "post": {
        "operationId": "archiveOwnerVenue",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Archive a venue catalog entry",
        "tags": [
          "venues"
        ]
      }
    },
    "/v1/venues/{id}/restore": {
      "post": {
        "operationId": "restoreOwnerVenue",
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueEnvelopeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "telegramMiniApp": []
          }
        ],
        "summary": "Restore a venue catalog entry",
        "tags": [
          "venues"
        ]
      }
    }
  },
  "info": {
    "title": "Football Bot API",
    "description": "Owner-facing REST API for the Football Telegram Mini App.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "telegramMiniApp": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Telegram-Init-Data",
        "description": "Raw Telegram.WebApp.initData. The server validates its HMAC and owner identity."
      }
    },
    "schemas": {
      "OwnerIdentityResponseDto": {
        "type": "object",
        "properties": {
          "telegramUserId": {
            "type": "string",
            "example": "123456789"
          }
        },
        "required": [
          "telegramUserId"
        ]
      },
      "GroupConfigurationResponseDto": {
        "type": "object",
        "properties": {
          "telegramChatId": {
            "type": "string",
            "example": "-1001234567890"
          },
          "generalTopicId": {
            "type": "string",
            "example": "1"
          },
          "weatherTopicId": {
            "type": "string",
            "example": "42"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/Minsk"
          }
        },
        "required": [
          "telegramChatId",
          "generalTopicId",
          "weatherTopicId",
          "timezone"
        ]
      },
      "MatchScheduleResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-08-03"
          },
          "time": {
            "type": "string",
            "example": "20:00"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/Minsk"
          }
        },
        "required": [
          "date",
          "time",
          "timezone"
        ]
      },
      "VenueBookingContactResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VenueResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "15"
          },
          "name": {
            "type": "string"
          },
          "mapUrl": {
            "type": "string",
            "format": "uri"
          },
          "venueType": {
            "type": "string",
            "enum": [
              "outdoor",
              "indoor"
            ]
          },
          "bookingContacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueBookingContactResponseDto"
            }
          },
          "websiteUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "mapUrl",
          "venueType",
          "bookingContacts",
          "websiteUrl",
          "archivedAt",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "PublicCardResponseDto": {
        "type": "object",
        "properties": {
          "publicationState": {
            "type": "string",
            "enum": [
              "not_published",
              "pending",
              "published",
              "uncertain",
              "failed",
              "deleted"
            ]
          },
          "telegramMessageId": {
            "type": "string",
            "nullable": true
          },
          "publicationAttemptedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "lastError": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "publicationState",
          "telegramMessageId",
          "publicationAttemptedAt",
          "lastError"
        ]
      },
      "MatchResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "42"
          },
          "chatId": {
            "type": "string",
            "example": "-1001234567890"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "schedule": {
            "$ref": "#/components/schemas/MatchScheduleResponseDto"
          },
          "venue": {
            "$ref": "#/components/schemas/VenueResponseDto"
          },
          "fieldPriceRubles": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "creatorTelegramUserId": {
            "type": "string",
            "example": "123456789"
          },
          "deletionRequestedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publicCard": {
            "$ref": "#/components/schemas/PublicCardResponseDto"
          }
        },
        "required": [
          "id",
          "chatId",
          "scheduledAt",
          "schedule",
          "venue",
          "fieldPriceRubles",
          "version",
          "creatorTelegramUserId",
          "deletionRequestedAt",
          "createdAt",
          "updatedAt",
          "publicCard"
        ]
      },
      "BootstrapResponseDto": {
        "type": "object",
        "properties": {
          "owner": {
            "$ref": "#/components/schemas/OwnerIdentityResponseDto"
          },
          "group": {
            "$ref": "#/components/schemas/GroupConfigurationResponseDto"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchResponseDto"
            }
          }
        },
        "required": [
          "owner",
          "group",
          "matches"
        ]
      },
      "MatchListResponseDto": {
        "type": "object",
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchResponseDto"
            }
          }
        },
        "required": [
          "matches"
        ]
      },
      "MatchEnvelopeResponseDto": {
        "type": "object",
        "properties": {
          "match": {
            "$ref": "#/components/schemas/MatchResponseDto"
          }
        },
        "required": [
          "match"
        ]
      },
      "MatchCreateDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-08-03",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "time": {
            "type": "string",
            "example": "20:00",
            "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
          },
          "venueId": {
            "type": "string",
            "example": "15",
            "pattern": "^[1-9]\\d*$"
          },
          "fieldPriceRubles": {
            "type": "number",
            "nullable": true,
            "example": 25,
            "minimum": 0
          }
        },
        "required": [
          "date",
          "time",
          "venueId"
        ]
      },
      "PatchMatchDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-08-03",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "time": {
            "type": "string",
            "example": "20:00",
            "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
          },
          "venueId": {
            "type": "string",
            "example": "15",
            "pattern": "^[1-9]\\d*$"
          },
          "fieldPriceRubles": {
            "type": "number",
            "nullable": true,
            "example": 25,
            "minimum": 0
          }
        }
      },
      "WeatherCurrentResponseDto": {
        "type": "object",
        "properties": {
          "sent": {
            "type": "boolean",
            "example": true
          },
          "observedAt": {
            "type": "string",
            "example": "2026-08-10T11:00"
          }
        },
        "required": [
          "sent",
          "observedAt"
        ]
      },
      "VenueListResponseDto": {
        "type": "object",
        "properties": {
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueResponseDto"
            }
          }
        },
        "required": [
          "venues"
        ]
      },
      "VenueBookingContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Администратор"
          },
          "phone": {
            "type": "string",
            "example": "+375 29 123-45-67"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VenueCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "BOX365 Октябрьская"
          },
          "mapUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://maps.google.com/?q=53.9,27.56"
          },
          "venueType": {
            "type": "string",
            "enum": [
              "outdoor",
              "indoor"
            ],
            "example": "indoor"
          },
          "bookingContacts": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueBookingContactDto"
            }
          },
          "websiteUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          }
        },
        "required": [
          "name",
          "mapUrl",
          "venueType"
        ]
      },
      "VenueEnvelopeResponseDto": {
        "type": "object",
        "properties": {
          "venue": {
            "$ref": "#/components/schemas/VenueResponseDto"
          }
        },
        "required": [
          "venue"
        ]
      },
      "VenueUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "BOX365 Октябрьская"
          },
          "mapUrl": {
            "type": "string",
            "format": "uri"
          },
          "venueType": {
            "type": "string",
            "enum": [
              "outdoor",
              "indoor"
            ]
          },
          "bookingContacts": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueBookingContactDto"
            }
          },
          "websiteUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          }
        }
      }
    }
  }
}