# packages/api-client/src/generated/model/matchCreateDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Football Bot API
 * Owner-facing REST API for the Football Telegram Mini App.
 * OpenAPI spec version: 1.0.0
 */

export interface MatchCreateDto {
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  /** @pattern ^(?:[01]\d|2[0-3]):[0-5]\d$ */
  time: string;
  /** @pattern ^[1-9]\d*$ */
  venueId: string;
  /**
   * @minimum 0
   * @nullable
   */
  fieldPriceRubles?: number | null;
}