# packages/api-client/src/generated/model/matchResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Football Bot API
 * Owner-facing REST API for the Football Telegram Mini App.
 * OpenAPI spec version: 1.0.0
 */
import type { MatchScheduleResponseDto } from './matchScheduleResponseDto.js';
import type { VenueResponseDto } from './venueResponseDto.js';
import type { PublicCardResponseDto } from './publicCardResponseDto.js';

export interface MatchResponseDto {
  id: string;
  chatId: string;
  scheduledAt: string;
  schedule: MatchScheduleResponseDto;
  venue: VenueResponseDto;
  /**
   * @minimum 0
   * @nullable
   */
  fieldPriceRubles: number | null;
  /** @minimum 1 */
  version: number;
  creatorTelegramUserId: string;
  /** @nullable */
  deletionRequestedAt: string | null;
  createdAt: string;
  updatedAt: string;
  publicCard: PublicCardResponseDto;
}