# packages/api-client/src/generated/model/publicCardResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Football Bot API
 * Owner-facing REST API for the Football Telegram Mini App.
 * OpenAPI spec version: 1.0.0
 */
import type { PublicCardResponseDtoPublicationState } from './publicCardResponseDtoPublicationState.js';

export interface PublicCardResponseDto {
  publicationState: PublicCardResponseDtoPublicationState;
  /** @nullable */
  telegramMessageId: string | null;
  /** @nullable */
  publicationAttemptedAt: string | null;
  /** @nullable */
  lastError: string | null;
}