# packages/api-client/src/generated/model/venueBookingContactDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Football Bot API
 * Owner-facing REST API for the Football Telegram Mini App.
 * OpenAPI spec version: 1.0.0
 */

export interface VenueBookingContactDto {
  name?: string;
  phone: string;
}